+++ console/src/components/Services/Data/Schema/Actions.js
@@ -5,9 +5,11 @@ import { getConfirmation } from '../../../Common/utils/jsUtils';
 import { dataSource } from '../../../../dataSources';
 import { getRunSqlQuery } from '../../../Common/utils/v1QueryUtils';
 import Migration from '../../../../utils/migration/Migration';
+import { setSidebarLoading } from '../DataSubSidebar';
 
 export const createNewSchema = (schemaName, successCb, errorCb) => {
   return (dispatch, getState) => {
+    dispatch(setSidebarLoading(true));
     const source = getState().tables.currentDataSource;
     if (!gqlPattern.test(schemaName)) {
       return dispatch(
@@ -34,12 +36,14 @@ export const createNewSchema = (schemaName, successCb, errorCb) => {
         if (successCb) {
           successCb();
         }
+        dispatch(setSidebarLoading(false));
       });
     };
     const customOnError = () => {
       if (errorCb) {
         errorCb();
       }
+      dispatch(setSidebarLoading(false));
     };
 
     makeMigrationCall(
@@ -59,6 +63,7 @@ export const createNewSchema = (schemaName, successCb, errorCb) => {
 
 export const deleteCurrentSchema = (successCb, errorCb) => {
   return (dispatch, getState) => {
+    dispatch(setSidebarLoading(true));
     const { currentSchema, currentDataSource } = getState().tables;
 
     const confirmMessage = `This will permanently delete schema "${currentSchema}" from the database`;
@@ -79,12 +84,16 @@ export const deleteCurrentSchema = (successCb, errorCb) => {
     const errorMsg = 'Error dropping schema';
 
     const customOnSuccess = () => {
-      dispatch(fetchSchemaList()).then(successCb);
+      dispatch(fetchSchemaList()).then(() => {
+        dispatch(setSidebarLoading(false));
+        successCb();
+      });
     };
     const customOnError = () => {
       if (errorCb) {
         errorCb();
       }
+      dispatch(setSidebarLoading(false));
     };
 
     makeMigrationCall(